01-resumelens-ai/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
cryptography==46.0.5
exceptiongroup==1.3.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.6
pyarrow==21.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==15.0.1
